% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Distribution_DeletEffects.R
\name{delet.muts.over.landscape}
\alias{delet.muts.over.landscape}
\title{Look at numbers of deleterious mutations across the landscape}
\usage{
delet.muts.over.landscape(del.file, patches.x, patches.y, num.loci,
  count.type = "total")
}
\arguments{
\item{del.file}{The file containing deleterious loci output from Nemo.}

\item{patches.x}{The number of patches on the landscape along the x-axis (parallel to expansion).}

\item{patches.y}{The number of patches on the landscape along the y-axis (perpendicular to expansion).}

\item{num.loci}{The number of deleterious loci that were simulated in that file.}

\item{count.type}{Whether to count homozygous, heterozygous, or total number of mutations.}
}
\value{
Creates a plot over the landscape (heat map style) for the mean number of deleterious mutations per individual in a patch.
}
\description{
Takes an output file from Nemo of deleterious loci genotypes, and plots the mean number of deleterious mutations per individual in a patch over the landscape.
}
\author{
Kimberly J Gilbert
}
\references{
\href{http://nemo2.sourceforge.net/index.html}{Nemo} is created and maintained by Fred Guillaume. The manual and source files are available online.
}

