% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Distribution_DeletEffects.R
\name{dist.delet.effects}
\alias{dist.delet.effects}
\title{Examine effect size distribution of deleterious loci from Nemo}
\usage{
dist.delet.effects(file, num.loci, xlim.ho = c(0, 1), xlim.he = c(0, 1))
}
\arguments{
\item{file}{The file containing deleterious loci output from Nemo.}

\item{num.loci}{The number of deleterious loci that were simulated in that file.}

\item{xlim.ho}{The limits of the x-axis for the homozygous effect distribution, default is from 0 to 1.}

\item{xlim.he}{The limits of the x-axis for the heterozygous effect distribution, default is from 0 to 1.}
}
\value{
Creates a plot of two histograms showing the distributions.
}
\description{
Takes an output file from Nemo of deleterious loci genotypes, and plots the distribution of both the homozygous and heterozygous effects of these loci.
}
\author{
Kimberly J Gilbert
}
\references{
\href{http://nemo2.sourceforge.net/index.html}{Nemo} is created and maintained by Fred Guillaume. The manual and source files are available online.
}

