% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimResults_Pop.R
\name{fit.results.landscape}
\alias{fit.results.landscape}
\title{Look at landscape fitness from .fit files}
\usage{
fit.results.landscape(input.fit.file, patches.x, patches.y,
  fitness.type = "total")
}
\arguments{
\item{input.fit.file}{The .fit file output from Nemo.}

\item{patches.x}{The number of patches on the landscape along the x-axis (parallel to expansion)}

\item{patches.y}{The number of patches on the landscape along the y-axis (perpendicular to expansion)}

\item{fitness.type}{Which component of fitness to plot: for the quantitative trait, for deleterious mutations, or total fitness.}
}
\value{
Creates four plots of simulation results, showing mean fitness, total number of adults in the metapopulation, mean phenotypic value, and variances across generations.
}
\description{
Takes an "ind_seln" output file (with quanti and delet fitness values per individual) and plots fitness over the landscape in several ways.
}
\author{
Kimberly J Gilbert
}
\references{
\href{http://nemo2.sourceforge.net/index.html}{Nemo} is created and maintained by Fred Guillaume. The manual and source files are available online.
}

