% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aNEMOne.R
\name{make.input}
\alias{make.input}
\title{Create Nemo input file}
\usage{
make.input(run.mode = "overwrite", random.seed = "12345",
  log.file = "logfile.log", root.dir = NULL, filename = NULL,
  reps = NULL, gens = NULL, num.patches = NULL, patch.capacity = NULL,
  patch.capacity.fem = NULL, patch.capacity.mal = NULL, cap.temp = FALSE,
  LCE.order = NULL, mating.system = NULL, mean.fec = NULL,
  seln.trait = NULL, seln.model = NULL)
}
\arguments{
\item{run.mode}{}

\item{random.seed}{Set the random seed for the run, default is 12345.}

\item{log.file}{Name of the logfile to be output.}

\item{root.dir}{}

\item{filename}{}

\item{reps}{}

\item{gens}{}

\item{num.patches}{}

\item{patch.capacity}{}

\item{patch.capacity.fem}{}

\item{patch.capacity.mal}{}

\item{cap.temp}{}

\item{LCE.order}{}

\item{mating.system}{}

\item{mean.fec}{}

\item{seln.trait}{}

\item{seln.model}{}
}
\value{
Write the file specified into the working directory.
}
\description{
Compares the QST of a single phenotypic trait to the mean FST of series of marker loci.
It calculates the distribution of QST - FST under a model assuming neutrality of both
the phenotypic trait and the genetic markers from which FST is estimated.
Returns the simulated estimates of Qst - Fst under neutrality following the procedure
described in Gilbert and Whitlock (2014) and Whitlock & Guillaume (2009). Also returns
the simulated estimates of Fst and Qst used to compute the null distribution.
}
\examples{
# population
cap <- patch.cap(6,1000, 0,2)
num <- 1002

# simulation components
life.cycles <- c("breed", "disperse", "selection", "aging")


# mating

breed.kernel <- make.kernel.and.matrix(cell.size=50, land.x=1000, land.y=2000, dist.mean=0, dist.sd=25, breed.window=TRUE, two.kernels=FALSE, second.dist.mean=0, second.dist.sd=NULL)


# dispersal

disp.kernel <- make.kernel.and.matrix <- function(cell.size=50, land.x=1000, land.y=2000, dist.mean=0, dist.sd=25, breed.window=FALSE, two.kernels=FALSE, second.dist.mean=0, second.dist.sd=NULL)


make.input(root.dir="test", filename="test2",
	reps=10, gens=100, num.patches=num, patch.capacity=cap,
	LCE.order= life.cycles, mating.system=1, mean.fec=5
	)
}
\author{
Kimberly J Gilbert
}

