% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/KernelsAndConnectivityMatrices.R
\name{make.kernel.and.matrix}
\alias{make.kernel.and.matrix}
\title{Create dispersal or breeding window kernels and connectivity matrices}
\usage{
make.kernel.and.matrix(cell.size, horizontal.land, vertical.land,
  dist.mean = 0, dist.sd, breed.window = FALSE, two.kernels = FALSE,
  second.dist.mean = 0, second.dist.sd = NULL)
}
\arguments{
\item{cell.size}{The size in unites distance of a single patch (or cell), which based on the parameters given below will influence how large the landscape is and how many cells are within one unit sigma.}

\item{horizontal.land}{The size of the landscape in the horizontal direction. Units are the same as cell.size.}

\item{vertical.land}{The size of the landscape in the vertical direction. Units are the same as cell.size.}

\item{dist.mean}{The mean distance for the kernel's distribution. Default is zero (i.e. most dispersal occurs to the natal patch).}

\item{dist.sd}{Sigma (one standard deviation) for the distribution of dispersal distances.}

\item{breed.window}{A boolean parameter with default FALSE. Change to true if making a breeding window and not a dispersal kernel. This changes the ID of the patch that is at the absorbing boundary of the landscape.}

\item{two.kernels}{A boolean parameter with default FALSE. If two distributions are being summed to create the kernel, this should be TRUE.}

\item{second.dist.mean}{The mean of the second distribution being summed with the first distribution. The default is zero, as above, but should match the first distribution's mean if that is ever changed from zero.}

\item{second.dist.sd}{Sigma (one standard deviation) for the distribution of dispersal distances of the second distribution.}
}
\value{
Returns the array of probabilities for the dispersal or breeding window kernels and prints the connectivity matrix to a file.
}
\description{
Automatically create the dispersal and breeding window probability kernels and connectivity matrices of patch IDs based on the provided cell size, number of cells in the x and y directions on the landscape, and the mean and standard deviation of the dispersal kernel. Additionally, two distributions are able to be summed for creating the dispersal kernel.
}
\examples{
make.kernel.and.matrix(cell.size=50, horizontal.land=1000, vertical.land=2000, dist.mean=0,
  dist.sd=25, breed.window=TRUE, two.kernels=FALSE, second.dist.mean=0, second.dist.sd=NULL)
}
\author{
Kimberly J Gilbert
}

