% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aNEMOne.R
\name{make.nemo.input}
\alias{make.nemo.input}
\title{Create Nemo input file for Nemo 2.3.46 official version, i.e. no breeding window or modifications from Fred's code.}
\usage{
make.nemo.input(run.mode = "overwrite", random.seed = "12345",
  log.file = "logfile.log", root.dir = NULL, filename = NULL,
  reps = NULL, gens = NULL, num.patches = NULL, patch.capacity = NULL,
  patch.capacity.fem = NULL, patch.capacity.mal = NULL, cap.temp = FALSE,
  LCE.order = NULL, mating.system = NULL, mating.proportion = NULL,
  mean.fec = NULL, large.kernels = FALSE, dispersal.kernel = NULL,
  dispersal.connectivity.matrix = NULL, seln.trait = NULL,
  seln.model = NULL, seln.fitness.model = "absolute", seln.var = NULL,
  seln.trait.dim = 1, seln.local.optima = NULL,
  seln.optima.rate.change = NULL, quanti.init = NULL,
  num.quanti.traits = 1, num.quanti.loci = NULL, quanti.mut.rate = NULL,
  quanti.mut.var = NULL, quanti.recomb.rate = 0.5, quanti.init.model = 1,
  quanti.env.var = 1, num.ntrl.loci = NULL, num.ntrl.alleles = NULL,
  ntrl.mut.rate = NULL, ntrl.recomb.rate = 0.5, ntrl.mut.model = NULL,
  ntrl.init.model = NULL, num.delet.loci = NULL, delet.mut.rate = NULL,
  delet.mut.model = 1, delet.recomb.rate = 0.5, delet.init.freq = 0,
  delet.effects.dist = "gamma", delet.effects.mean = NULL,
  delet.effects.dist.param1 = NULL, delet.effects.dist.param2 = NULL,
  delet.dominance.mean = NULL, delet.fitness.model = 1, save.ntrl = NULL,
  save.quanti = NULL, save.delet = NULL, save.stats = NULL,
  save.seln.inds = NULL, stats = NULL)
}
\arguments{
\item{run.mode}{The mode to run the simulation in; default is "overwrite", for other options see the Nemo manual.}

\item{random.seed}{Set the random seed for the run, default is 12345.}

\item{log.file}{Name of the logfile to be output.}

\item{root.dir}{Root directory to put run outputs.}

\item{filename}{Base name of the simulation file.}

\item{reps}{Number of replicates to perform.}

\item{gens}{Number of generations to run the simulation.}

\item{num.patches}{Number of patches on the landscape.}

\item{patch.capacity}{The carrying capacity of each patch, can be specified by an integer value or an array of length num.patches.}

\item{patch.capacity.fem}{If sex-specific carrying capacity, this is the female carrying capacity per patch.}

\item{patch.capacity.mal}{If sex-specific carrying capacity, this is the male carrying capacity per patch.}

\item{cap.temp}{If carrying capacity changes at a given generation, this is the generation at which that occurs.}

\item{LCE.order}{The order of life cycle events in the simulation. Should be given as multiple character strings, in order.}

\item{mating.system}{See the nemo manual for details, random mating is specified by a '1'.}

\item{mating.proportion}{See the nemo manual for details, sets the proportion of non-random mating.}

\item{mean.fec}{Mean fecundity per mother, may also be set to be patch-specific with an array. By default follows a Poisson distribution, see the Nemo manual for details.}

\item{large.kernels}{Set to true if using a dispersal or breeding kernel of large size, default is false. When true, it will not hold the full matrix in R's memory and instead send to the command line to concatenate files.}

\item{dispersal.kernel}{The matrix of dispersal probabilities for forward migration. (parameter in Nemo is dispersal_reduced_matrix)}

\item{dispersal.connectivity.matrix}{The connectivity matrix of patches matched to the dispersal kernel.}

\item{seln.trait}{The trait(s) specified to be under selection. Either "quant" or "delet" for quantitative or deleterious traits.}

\item{seln.model}{The model of selection to use. See Nemo manual section 4.7 for all details on selection.}

\item{seln.fitness.model}{Options are absolute, relative_global, and relative_local. Default is absolute. See Nemo manual section 4.7 for all details on selection.}

\item{seln.var}{Variance in selection.}

\item{seln.trait.dim}{See Nemo manual section 4.7 for all details on selection. Default is 1.}

\item{seln.local.optima}{Optimal trait value per patch or for the entire landscape}

\item{seln.optima.rate.change}{Rate of change per generation for the optimal trait value per patch or for the entire landscape}

\item{quanti.init}{The initial QTL value to be set; see Nemo manual section 5.3 for all details on quantitative traits.}

\item{num.quanti.traits}{How many QTL traits to model, default is 1.}

\item{num.quanti.loci}{How many loci underly the quantitative trait.}

\item{quanti.mut.rate}{Mutation rate per QTL.}

\item{quanti.mut.var}{Variance in mutation rate per QTL.}

\item{quanti.recomb.rate}{Recombination distances among QTL on a random genetic map.}

\item{quanti.init.model}{See Nemo manual section 5.3 for all details on quantitative traits. Default is 1.}

\item{quanti.env.var}{Environmental variance, default is 1.}

\item{num.ntrl.loci}{The number of neutral loci to simulate. See Nemo manual section 5.2 for all details on neutral markers.}

\item{num.ntrl.alleles}{The number of alleles per neutral locus.}

\item{ntrl.mut.rate}{The mutation rate per neutral locus.}

\item{ntrl.recomb.rate}{The recombination rate among neutral loci.}

\item{ntrl.mut.model}{The mutation model for neutral loci: 1 = single step, 2 = K allele model. See Nemo manual section 5.2 for all details on neutral markers.}

\item{ntrl.init.model}{How to initiate the neutral allele frequencies: 0 = no initial variance, 1 = maximum initial variance. See Nemo manual section 5.2 for all details on neutral markers.}

\item{num.delet.loci}{The number of deleterious loci to simulate. See section 5.4 of the Nemo manual for details on deleterious mutation parameters.}

\item{delet.mut.rate}{The mutation rate per deleterious locus.}

\item{delet.mut.model}{Default is 1, meaning a new mutation locus drawn irrespective of presence of existing mutation (model 2 skips if locus has already mutated).}

\item{delet.recomb.rate}{Recombination distances among deleterious loci on a random genetic map.}

\item{delet.init.freq}{Initial frequency of deleterious mutations, default is 0.}

\item{delet.effects.dist}{The shape of the distribution used to model deleterious effects, default is gamma. (other options are constant, exponential, and lognormal)}

\item{delet.effects.mean}{The mean of the distribution of deleterious effects.}

\item{delet.effects.dist.param1}{The shape of the gamma distribution.}

\item{delet.effects.dist.param2}{The scale of the gamma distribution; not needed when mean and shape are specified.}

\item{delet.dominance.mean}{Mean value of h, the dominance coefficient.}

\item{delet.fitness.model}{The fitness model used to compute individual viability, default is multiplicative (1), additive (2) is the other option.}

\item{save.ntrl}{How often to save the neutral marker genotype files. Files are automatically output to subdirectory "ntrl_geno". If not present, no neutral genotype output is saved.}

\item{save.quanti}{How often to save the quantitative trait genotype files. Files are automatically output to subdirectory "quanti_geno". If not present, no quantitative genotype output is saved.}

\item{save.delet}{How often to save the deleterious mutation files. Files are automatically output to subdirectory "delet_geno". If not present, no deleterious output is saved.}

\item{save.stats}{How often to save the values of the parameters defined by "stats". Files are automatically output to subdirectory "stats". *If this paramter is not present, no stats are output, even if "stats" is defined.}

\item{save.seln.inds}{How often to save the data from "selection_output". Files are automatically output to subdirectory "ind_seln". *If this paramter is not present, no selection data are output.}

\item{stats}{Population and simlation parameters to return, see Nemo manual section 7 "Output Statistics".}
}
\value{
Write the file specified into the working directory.
}
\description{
Updated for version 2.3.46. Makes an input file with deleterious loci and QTL of type .ini for \href{http://nemo2.sourceforge.net/index.html}{Nemo}, an individual-based, forward time simulation program created and maintained by Fred Guillaume.
}
\author{
Kimberly J Gilbert
}
\references{
\href{http://nemo2.sourceforge.net/index.html}{Nemo} is created and maintained by Fred Guillaume. The manual and source files are available online.
}

