% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WestGridFunctions.R
\name{make.pbs}
\alias{make.pbs}
\title{Make Westgrid PBS files}
\usage{
make.pbs(filename, input.directory = NULL, nemo.directory = NULL, RAM,
  walltime = NULL, westgrid.server = NULL)
}
\arguments{
\item{filename}{The full name of the .ini file for which the PBS file is being created.}

\item{input.directory}{The directory on the Westgrid server which holds the nemo .ini file. Defaults are set for each server.}

\item{nemo.directory}{The directory on the Westgrid server from which the nemo executable is called. Defaults are set for each server.}

\item{RAM}{The amount of RAM the run will be allowed to use, there is no default, and must include gb, mb, or kb in specification, e.g. "6gb".}

\item{walltime}{Default is set at max for specified server (except bugaboo, which is set to default at 500 hours), but other lengths can be specified here in units "100:00:00" hours:minutes:seconds.}

\item{westgrid.server}{Which westgrid server the PBS file is meant for. Options are "grex", "orcinus" or "bugaboo" currently.}
}
\value{
Creates PBS files from a .ini file for running Nemo on Westgrid's Orcinus, Bugaboo, or Grex servers. *IMPORTANT NOTE* Must have files "MiddleLine.txt", "GrexLine.txt", and "LastLine.txt" in the same directory for make.pbs to work properly, as it concatenates text from those files into the final PBS file.
}
\description{
Take an individual Nemo .ini file and make a Westgrid PBS file for either orcinus, bugaboo, or grex servers.
}
\author{
Kimberly J Gilbert
}
\references{
\href{http://nemo2.sourceforge.net/index.html}{Nemo} is created and maintained by Fred Guillaume. The manual and source files are available online.
}

