% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PatchCapacity.R
\name{patch.cap}
\alias{patch.cap}
\title{Create array of patch capacities}
\usage{
patch.cap(K1, num.K1, K2 = NULL, num.K2 = NULL, temp.gen = NULL)
}
\arguments{
\item{K1}{The carrying capacity for patches in the array. If there will be two different carrying capacities, this is for the first set of patches in the array.}

\item{num.K1}{The number of patches having carrying capacity, K1.}

\item{K2}{The carrying capacity for the second set of patches in the array. Only needed if there is more than one K for all patches.}

\item{num.K2}{The number of patches having carrying capacity, K2. Only needed if there is more than one K for all patches.}

\item{temp.gen}{If temporal capacities are desired, this is the generation at which the provided capacities will be set.}
}
\value{
Returns the array of patch capacities that can then be provided to the 'make.input' function for the final input file.
}
\description{
Make the linear array pf patch capacities to be used in the Nemo .init file.
}
\examples{
patch.cap(K1=10, num.K1=12, K2=0, num.K2=4)
}
\author{
Kimberly J Gilbert
}
\references{
Gilbert KJ
}

