% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimResults_Pop.R
\name{plot1D.results}
\alias{plot1D.results}
\title{Look at 1-D simulation results over landscape}
\usage{
plot1D.results(input.fit.file, input.quanti.file = NULL,
  input.del.file = NULL, patches.x, patches.y, slope.opt = NULL,
  opt.rate.change = NULL, delay.env.change = NULL, generation, del.loci,
  xlimits = NULL, plot.legend = TRUE, Vs = NULL, quantiles = c(0.5,
  0.1))
}
\arguments{
\item{input.fit.file}{The .fit file output from Nemo.}

\item{input.quanti.file}{The .quanti file output from Nemo which contains genotypes for the quantitative trait.}

\item{input.del.file}{The .del file output from Nemo which contains deleterious mutation genotypes.}

\item{patches.x}{The number of patches on the landscape along the x-axis (parallel to expansion)}

\item{patches.y}{The number of patches on the landscape along the y-axis (perpendicular to expansion)}

\item{slope.opt}{The slope of the environmental optimum defined in the input file.}

\item{opt.rate.change}{The rate of change of the environment per generation defined in the input file.}

\item{delay.env.change}{The gneeration at which the environment starts changing, if it is not immediate.}

\item{generation}{The current generation being plotted.}

\item{del.loci}{The number of deleterious loci simulated.}

\item{xlimits}{A smaller range on the x-axis to plot, if desired.}

\item{plot.legend}{Whether to include a legend or not on the fitness and deleterious mutation number plots.}

\item{Vs}{The value of selection variance.}

\item{quantiles}{The value of the two quantiles to plot for the selection distribution around the optimum.}
}
\value{
Creates 1-D plots over the landscape for population size, fitness, numbers of deleterious mutations, and environmental optimum match to the quantitative trait.
}
\description{
Makes 1-D plots over the landscape for population size, fitness, numbers of deleterious mutations, and environmental optimum match to the quantitative trait. Simulations must have been in 1 dimension for current implementation.
}
\author{
Kimberly J Gilbert
}
\references{
\href{http://nemo2.sourceforge.net/index.html}{Nemo} is created and maintained by Fred Guillaume. The manual and source files are available online.
}

