% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aNEMOne.R
\name{read.output}
\alias{read.output}
\title{Analyze a Nemo stat output file}
\usage{
read.output(filename, params.set)
}
\arguments{
\item{filename}{The name and path to the file to be analyzed.}

\item{params.set}{A lsit of parameters that were set to be included in the stat output file}
}
\value{
Returns either a concise list of a subset of results or a full list with all possible results. Both output
}
\description{
Analyze output from Nemo specified by the stat options from a Nemo .init file
}
\details{

}
\author{
Kimberly J Gilbert
}

