% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimResults_Pop.R
\name{sim.results.pop}
\alias{sim.results.pop}
\title{Examine some population parameters from Nemo stat output}
\usage{
sim.results.pop(input.stats.file, which.rep = NULL, par.size = c(2, 2),
  stats.to.plot = c("fitness.mean", "adlt.nbr", "adlt.q1", "adlt.q1.Va"))
}
\arguments{
\item{input.stats.file}{The stat file output from Nemo.}

\item{which.rep}{If there are stats from multiple replicates, state which one to examine here, otherwise leave as NULL.}

\item{par.size}{Create the layout/number of plotting windows for making the desired output plots.}

\item{stats.to.plot}{List the stats to be plotted.}
}
\value{
Creates four plots of simulation results, showing mean fitness, total number of adults in the metapopulation, mean phenotypic value, and variances across generations.
}
\description{
Takes an output file from Nemo of deleterious loci genotypes, and plots the distribution of both the homozygous and heterozygous effects of these loci.
}
\author{
Kimberly J Gilbert
}
\references{
\href{http://nemo2.sourceforge.net/index.html}{Nemo} is created and maintained by Fred Guillaume. The manual and source files are available online.
}

