% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SelectionLandscape.R
\name{step.landscape}
\alias{step.landscape}
\title{Create the landscape of optimal phenotypic values that change at an equal and constant pace over space}
\usage{
step.landscape(horizontal.patches, vertical.patches, step.width, scale.to = 1)
}
\arguments{
\item{horizontal.patches}{Number of patches in the horizontal, x direction.}

\item{vertical.patches}{Number of patches in the vertical, y direction.}

\item{step.width}{The number of patches that horizontally make up the width of "steps" of equal optimal phenotype to create the landscape. At most can equal the number of vertical columns in the landscape to create a uniform landscape, or at minimum can equal 1 for the landscape to change every single column of patches. This number must equally divide into the total number of horizontal patches.}

\item{scale.to}{Value to scale down the magnitude of change in optima over the total landscape. Default is 1, which scales the total change to approximately -1 to 1 across the full landscape, whereas, e.g. a value of 5 allows a higher degree of change across the landscape, approximately -5 to 5.}
}
\value{
Returns the value of the first column of patches at the leftmost end of the landscape. Also prints to a file the matrix of values for the landscape, to be uses as "selection_local_optima" in Nemo, and additionally visualizes the landscape.
}
\description{
Create the landscape of optimal phenotypic values over which the patches will exist. Size should match the number of cells in the x and y directions on the landscape. A figure is created to visualize the landscape as well as the matrix of values output to a file for use in the make.input function.
}
\examples{
step.landscape(horizontal.patches=100, vertical.patches=25, step.width=10)
}
\author{
Kimberly J Gilbert
}

