% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aNEMOne.R
\name{make.input}
\alias{make.input}
\title{Create Nemo input file}
\usage{
make.input(run.mode = "overwrite", random.seed = "12345",
  log.file = "logfile.log", root.dir = NULL, filename = NULL,
  reps = NULL, gens = NULL, num.patches = NULL, patch.capacity = NULL,
  patch.capacity.fem = NULL, patch.capacity.mal = NULL, cap.temp = FALSE,
  LCE.order = NULL, mating.system = NULL, mean.fec = NULL,
  seln.trait = NULL, seln.model = NULL)
}
\arguments{
\item{run.mode}{The mode to run the simulation in; default is "overwrite", for other options see the Nemo manual.}

\item{random.seed}{Set the random seed for the run, default is 12345.}

\item{log.file}{Name of the logfile to be output.}

\item{root.dir}{Root directory to put run outputs.}

\item{filename}{Base name of the simulation file.}

\item{reps}{Number of replicates to perform.}

\item{gens}{Number of generations to run the simulation.}

\item{num.patches}{Number of patches on the landscape.}

\item{patch.capacity}{The carrying capacity of each patch, can be specified by an integer value or an array of length num.patches.}

\item{patch.capacity.fem}{If sex-specific carrying capacity, this is the female carrying capacity per patch.}

\item{patch.capacity.mal}{If sex-specific carrying capacity, this is the male carrying capacity per patch.}

\item{cap.temp}{If carrying capacity changes at a given generation, this is the generation at which that occurs.}

\item{LCE.order}{The order of life cycle events in the simulation. Should be given as multiple character strings, in order.}

\item{mating.system}{See the nemo manual for details, random mating is specified by a '1'.}

\item{mean.fec}{Mean fecundity per mother, may also be set to be patch-specific with an array. By default follows a Poisson distribution, see the Nemo manual for details.}

\item{seln.trait}{The trait(s) specified to be under selection.}

\item{seln.model}{The model of selection to use.}
}
\value{
Write the file specified into the working directory.
}
\description{
Compares the QST of a single phenotypic trait to the mean FST of series of marker loci.
It calculates the distribution of QST - FST under a model assuming neutrality of both
the phenotypic trait and the genetic markers from which FST is estimated.
Returns the simulated estimates of Qst - Fst under neutrality following the procedure
described in Gilbert and Whitlock (2014) and Whitlock & Guillaume (2009). Also returns
the simulated estimates of Fst and Qst used to compute the null distribution.
}
\examples{
# population
cap <- patch.cap(6,1000, 0,2)
num <- 1002

# simulation components
life.cycles <- c("breed", "disperse", "selection", "aging")


# mating

breed.kernel <- make.kernel.and.matrix(cell.size=50, land.x=1000, land.y=2000, dist.mean=0, dist.sd=25, breed.window=TRUE, two.kernels=FALSE, second.dist.mean=0, second.dist.sd=NULL)


# dispersal

disp.kernel <- make.kernel.and.matrix <- function(cell.size=50, land.x=1000, land.y=2000, dist.mean=0, dist.sd=25, breed.window=FALSE, two.kernels=FALSE, second.dist.mean=0, second.dist.sd=NULL)


make.input(root.dir="test", filename="test2",
	reps=10, gens=100, num.patches=num, patch.capacity=cap,
	LCE.order= life.cycles, mating.system=1, mean.fec=5
	)
}
\author{
Kimberly J Gilbert
}

