% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimResults_Pop.R
\name{sim.results.pop}
\alias{sim.results.pop}
\title{Examine some population parameters from Nemo stat output}
\usage{
sim.results.pop(input, out.file = "SimResults.pdf", which.rep = NULL,
  add = FALSE)
}
\arguments{
\item{input}{The stat file output from Nemo.}

\item{out.file}{The name of the file to output the .pdf of figures to, currently not enabled in the function.}

\item{which.rep}{If there are stats from multiple replicates, state which one to examine here, otherwise leave as NULL.}

\item{add}{If adding to an existing plor, change to TRUE, otherwise leave at default of FALSE I think this will probably not work as is.}
}
\value{
Creates four plots of simulation results, showing mean fitness, total number of adults in the metapopulation, mean phenotypic value, and variances across generations.
}
\description{
Takes an output file from Nemo of deleterious loci genotypes, and plots the distribution of both the homozygous and heterozygous effects of these loci.
}
\author{
Kimberly J Gilbert
}
\references{
\href{http://nemo2.sourceforge.net/index.html}{Nemo} is created and maintained by Fred Guillaume. The manual and source files are available online.
}

